I changed my mind about the approach to this project
Delete existing code for a fresh start. Code I'm happy with is in mblcxx and
mbltrc this will be ported into new code, with some changes.

Enforcing a associated .cpp for every header inside the project itself seems
redundant when using tools like iwyu. The test file should have the
header as the first include, this together with the tooling should be
enough to prove that the header is self contained.

mblcxx was mostly ok except the remark above. mbltrc was developed with C++
objects on a shared objects boundaries, this was a mistake.

mqlast and mqlprs took the approach of parsing each production
individually but this didn't turn out too good due to the complexity of
the grammar. While mqlast will probably at some point exist in a similar
form soon, for now it doesn't make sense to maintain it, new approach to
parsing will be done by first tokenizing the input and then parsing it.

mdbsql was mostly gutted and waiting for parsing to be finished.

mclxmp, mdbsrv were mostly prototype code, while mdbnet should have
provided better abstractions around network and serialization.